Refactor geometry module as lineartransformations
